redmine_agent/requirements.txt
# シンプルな依存関係を指定
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.24.0
requests>=2.28.0,<2.32.0
python-dotenv>=1.0.0,<1.1.0
pydantic>=1.10.5,<2.0.0

# LLM機能
google-generativeai>=0.3.0
protobuf>=4.24.4

# スケジューリング機能
pytz>=2023.3

# 以下は使用する場合は別途インストールが必要
# line-bot-sdk>=3.1.0
# 注意: line-bot-sdkとaiohttpは現在のPythonバージョンと互換性の問題がある場合があります
